[EN] Make config throw errors and take defaults.
`config` is now a 2-arity function. It takes the key path and an
optional default value.

If no default is given and the key path does not exist in the config
file, an exception is thrown. I tried to make the exception informative
as to what was expected to exist and where. It contains the key path,
the current config file's value (after reading), and the filename of the
config file.

If a default is given and the key path does not exist, the default is
returned.

We discussed this at length on Slack. The main problem was tracking down
exceptions (often due to `nil`s) that were caused by missing or
misspelled config items. It was pretty unanimous that at least a warning
should be issued if a config value was looked for but it didn't exist.

The debate had two sides: one claiming that it should throw an exception
if the value does not exist and one claiming that this was drastic and
the code reading from the config should use idiomatic `(or (config :x)
default)` or `(if-let [val (config :x)]...` constructs.

This is an implementation of the former side. There is still a single
function with two arities. The choice is obvious: if you have a default,
put it, otherwise, don't. If you want to avoid the exception, put a
default (could be `nil`). Either way everything is explicit.

Finally, the tests were modified to use the new key path style (using
vectors). New tests were added to test the exception and default
behavior.