[EN] Use get-in and don't coerce resource uri to file.
It was pointed out by Chris that the 2 arg arity is just a `get-in`,
greatly simplifying that code.

He also pointed out that not all resource urls can be converted to
Files. When you're in a JAR, it's a different kind of URL. So we'll just
leave it as a URL and that will be that.